--- app/(tabs)/inbox/buying.tsx
@@ -1,9 +1,0 @@
-import { Label, ScreenContainer } from '@/components/atoms';
-
-export default function BuyingInboxScreen() {
-  return (
-    <ScreenContainer>
-      <Label>Buying Screen</Label>
-    </ScreenContainer>
-  );
-}
