--- app/(tabs)/inbox/my-bids.tsx
@@ -1,9 +1,0 @@
-import { Label, ScreenContainer } from '@/components/atoms';
-
-export default function MyBidsInboxScreen() {
-  return (
-    <ScreenContainer>
-      <Label>My Bids Screen</Label>
-    </ScreenContainer>
-  );
-}
